// src/screens/CoacheScreens/MainScreens/clients-goals/global-goasl-editing-screen/model/constructor-state.ts
import { GoalsResArray, GoalsResponseProps } from "../interface/interface";
import { Goals } from "./goals";

export interface StateInitial {
  type: "INITIAL";
  data: null;
  error: null;
}

export interface StateLoading {
  type: "LOADING";
  data: GoalsResArray | null;
  error: null;
}

export interface StateHasData {
  type: "HAS_DATA";
  data: Array<Goals>;
  error: null;
}

export interface StateEmpty {
  type: "EMPTY";
  data: Record<string, never>;
  error: null;
}

export interface StateError {
  type: "ERROR";
  data: null;
  error: Error;
}

export type ConstructorState =
  | StateInitial
  | StateLoading
  | StateHasData
  | StateEmpty
  | StateError;